DoRun: fix possible invalid read
If a string passed to DoRun is less than sizeof(size_t), then its content
can't be read as if its type is size_t*. DoRun splits the string to 3
parts: pre-aligned, algned and post-aligned. If there is no aligned part
of a string, then whole string is treated as pre- and post-aligned.

The error was caused by casting pre- and post-aligned parts to size_t*.
Now they are processed by code operating on char* (function SafeRunChunk).
(Actually pre-aligned part is allowed to be casted to size_t*, if aligned
part exists, ie. length of the string > size(size_t).)