Replace Features with Options from easy.h
This is intended for both support of `Pire::Regexp` class from
pire/easy.h and simpler enum-based interface to features and encodings.

`Pire::Option<...>` and all the `Pire::Options::Proxy` magic classes proved to
be too hard to wrap in Cython, so `Pire::Option` is replaced with integer
enum `OptionFlag`, and `Pire::Options` is represented literally as set of
`OptionFlag`'s, which is converted to `Pire::Options` where needed.

Mako provides a way to manage the set of available
options/`Feature`s/`Encoding`s: once a new option is needed, it should
suffice to add a corresponding `OptionSpec` to render.py and probably an
`#include` to options.h.mako.