Add SlowCapturingScanner (#51)
* Add SlowCapturingScanner

Add the new version of CapturingScanner based on NFA, which is more correct
than the previous one. Add new operators of 'non-greedy' repetitions, for example,
'+?', '??', '*?', which give a chance for users to tell, that the text, captured by
those repetitions, should be the shortest one.

Add c++11 flag to configure.ac.